refactor: changed file name professors.js to professor.js
